Mock functions in a package.

Mocking allows you to temporary replace the implementation of functions within a package, which useful for testing code that relies on functions that are slow, have unintended side effects or access resources that may not be available when testing.

Up until recently, such capability was offered via testthat::with_mock(), but with release of version 3.0.0 and introduction of edition 3, this was deprecated from testthat, leaving it to third party packages to replace this feature. This mocking implementation is powered by utils::assignInNamespace() and therefore, caveats outlined in the corresponding documentation apply here too.

with_mock(..., mock_env = pkg_env(), eval_env = parent.frame())

Arguments

...
Named parameters redefine mocked functions, unnamed parameters will be evaluated after mocking the functions

mock_env
The environment in which to patch the functions, defaults to the top-level environment. A string is interpreted as package name.

eval_env
Environment in which expressions passed as ... are evaluated, defaults to parent.frame().

Value

The result of the last unnamed argument passed as ... (evaluated in the environment passed as eval_env)

Examples

url <- "https://eu.httpbin.org/get?foo=123"
mok <- function(...) "mocked request"

with_mock(
  `curl::curl_fetch_memory` = mok,
  curl::curl_fetch_memory(url)
)

#> [1] "mocked request"

dl_fun <- function(x) curl::curl_fetch_memory(x)

with_mock(
  `curl::curl_fetch_memory` = mok,
  dl_fun(url)
)

#> [1] "mocked request"

json <- function(...) '["mocked request"]'

with_mock(
  `curl::curl` = json,
  jsonlite::fromJSON(url)
)

#> [1] "mocked request"

with_mock(
  `curl::curl` = json,
  parse_and_simplify = function(txt, ...) gsub('\\[?\\"\\]?', "", txt),
  jsonlite::fromJSON(url),
  mock_env = "jsonlite"
)

#> [1] "mocked request"